requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
pillow